Ensure fingerprint is consistent across execution for all task types
It was discovered in staging tests for thenpingme, that we were not
correctly calculating the fingerprint of job-based scheduled tasks.

It seems that the `spl_object_hash` and `spl_object_id` vary from start
to finish event due to the way closures are serialised by PHP.

As a result, remove them from the calculation, and add tests for each of
the supported scheduled task types to ensure that the fingerprint will
remain consistent between the start and end events being fired.

Fixes #3