Add the ability to override the configured project name
By default, thenping.me uses the configured `APP_NAME` (`app.name`)
value for setting the project name within the UI.

Whilst this means there is no configuration required to get started, it
does limit the ability to differentiate between deployed instances of
the same application for multiple environments or clients.

This is a backwards compatible change that introduces a
`thenpingme.project_name` configuration value that can be set with the
`THENPINGME_PROJECT_NAME` environment variable and falls back to
`app.name` that is set by the `APP_NAME` environment variable.

Fixes #26